chore: extract SimCloudFrontDistributionConfigurator
